Optimize code Navigator & Realm
