+++ app/src/main/java/com/fstyle/structure_android/data/source/RepositoryModule.java
@@ -1,9 +1,7 @@
 package com.fstyle.structure_android.data.source;
 
 import com.fstyle.structure_android.data.source.local.realm.RealmApi;
-import com.fstyle.structure_android.data.source.local.realm.RealmImpl;
 import com.fstyle.structure_android.utils.dagger.AppScope;
-
 import dagger.Module;
 import dagger.Provides;
 
@@ -17,6 +15,6 @@ public class RepositoryModule {
     @AppScope
     @Provides
     public RealmApi provideRealmApi() {
-        return new RealmImpl();
+        return new RealmApi();
     }
 }
