+++ app/src/main/java/com/fstyle/structure_android/data/source/local/realm/UserLocalDataSource.java
@@ -1,17 +1,13 @@
 package com.fstyle.structure_android.data.source.local.realm;
 
 import android.support.annotation.NonNull;
-
 import com.fstyle.structure_android.data.model.User;
 import com.fstyle.structure_android.data.source.UserDataSource;
-
-import java.util.List;
-
-import javax.inject.Inject;
-
 import io.realm.Realm;
 import io.realm.RealmObject;
 import io.realm.RealmResults;
+import java.util.List;
+import javax.inject.Inject;
 import rx.Observable;
 import rx.Subscriber;
 import rx.functions.Action2;
@@ -32,7 +28,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
     @Override
     public void openTransaction() {
         if (mRealmApi == null) {
-            mRealmApi = new RealmImpl();
+            mRealmApi = new RealmApi();
         }
     }
 
@@ -48,7 +44,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
 
     @Override
     public Observable<Void> insertUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
+        return mRealmApi.realmTransactionAsync(new Action2<Subscriber<? super Void>, Realm>() {
             @Override
             public void call(Subscriber<? super Void> subscriber, Realm realm) {
                 realm.insert(user);
@@ -59,7 +55,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
 
     @Override
     public Observable<Void> updateUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
+        return mRealmApi.realmTransactionAsync(new Action2<Subscriber<? super Void>, Realm>() {
             @Override
             public void call(Subscriber<? super Void> subscriber, Realm realm) {
                 realm.insertOrUpdate(user);
@@ -70,7 +66,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
 
     @Override
     public Observable<Void> deleteUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
+        return mRealmApi.realmTransactionAsync(new Action2<Subscriber<? super Void>, Realm>() {
             @Override
             public void call(Subscriber<? super Void> subscriber, Realm realm) {
                 RealmObject.deleteFromRealm(user);
@@ -81,7 +77,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
 
     @Override
     public Observable<Void> insertOrUpdateUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
+        return mRealmApi.realmTransactionAsync(new Action2<Subscriber<? super Void>, Realm>() {
             @Override
             public void call(Subscriber<? super Void> subscriber, Realm realm) {
                 realm.insertOrUpdate(user);
@@ -92,7 +88,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
 
     @Override
     public Observable<List<User>> getAllUser() {
-        return mRealmApi.realmOnNewThread(new Action2<Subscriber<? super List<User>>, Realm>() {
+        return mRealmApi.realmGet(new Action2<Subscriber<? super List<User>>, Realm>() {
             @Override
             public void call(Subscriber<? super List<User>> subscriber, Realm realm) {
                 RealmResults<User> users = realm.where(User.class).findAll();
@@ -105,7 +101,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
 
     @Override
     public Observable<User> getUserByUserLogin(final String userLogin) {
-        return mRealmApi.realmOnNewThread(new Action2<Subscriber<? super User>, Realm>() {
+        return mRealmApi.realmGet(new Action2<Subscriber<? super User>, Realm>() {
             @Override
             public void call(Subscriber<? super User> subscriber, Realm realm) {
                 User user = realm.where(User.class).equalTo("login", userLogin).findFirst();
